requirements.txt
flask 
flask-cors
python-dotenv
requests
pandas
python-dotenv
pymongo
